Style compomenent in my idea list jsx
