--- src/components/pages/mylist/MyIdeaList.scss
@@ -1,5 +1,0 @@
-.idea-list {
-  background: #eee;
-  padding: 32px;
-  margin-top: 32px;
-}
